backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
groq==0.4.2
langgraph>=0.0.26
langchain>=0.1.10
langchain-core>=0.1.25
langchain-groq>=0.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
python-docx==1.1.0
pytesseract==0.3.10
sentence-transformers>=2.3.0
huggingface-hub>=0.19.0
faiss-cpu==1.7.4
numpy==1.26.3
httpx==0.26.0
redis==5.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
